Refactored the Create CannonUtil Branch
Refactored the latest commit on this branch:
1. Removed CannonUtilTemp from OperateDrive Command.
2. Created OperateCannon Command
3. Updated RobotContainer to instantiate CannonUtilTemp and map XBox Buttons B and Y to activate the OperateCannon Command to shoot Barrels 1 and 2 respectively.
4. Demonstrated best practice for Command, Subsystem and SmartDashboard coding in the WPILibJ Commandbase Robot Framework.

This commit should be a teaching guide for students learning the framework we are using this year.

This code is a clean build on my local and should run as is on a robot if deployed from a correctly configured laptop for the FRC 2022 framework and the Driver Station software is also installed.