Merger Sparrow Drive Mode into the TSC2022 Project
The DriveSparrow project is not compliant with the FRC2022 framework and could not be merged.  It had to be recoded.  This needs careful testing before merging this branch to master.

Here is a high level of the changes:
1. OperateDrive - removed the need to pass in double values.  Subsystem can get them directly as needed.  Simplifies the command.
2. Constants - added a bunch
3. DriveMode - deleted ENUM class as it is not used anymore.
4. RobotContainer - moved the SendableChoosers there.  Best to keep them in one place.  Made the values Public Static so they can be referenced safely as needed in both Command and Subsystem classes.  Remove Toggle Drive Mode command button.
5. DriveUtil - Removed Toggle Drive method as drive type is now based on SendableChooser() and not an ENUM class.  Created helper classes setDriveControls and sparrowDrive()

This commit compiles clean so you should be able to deploy it to the robot from a laptop that is compliant with the FRC 2022 framework and has the current Driver Station software installed.